+++ src/noahtheduke/splint.clj
@@ -81,6 +81,7 @@
     noahtheduke.splint.rules.style.filter-vec-filterv
     noahtheduke.splint.rules.style.first-first
     noahtheduke.splint.rules.style.first-next
+    noahtheduke.splint.rules.style.is-eq-order
     noahtheduke.splint.rules.style.let-do
     noahtheduke.splint.rules.style.mapcat-apply-apply
     noahtheduke.splint.rules.style.mapcat-concat-map